// client/src/generated/sinks/log.ts
export interface LogEntry {
    /**
     * @generated from protobuf field: uint64 id = 1
     */
    id: bigint;
    /**
     * @generated from protobuf field: string message = 2
     */
    message: string;
    /**
     * @generated from protobuf field: int64 timestamp = 3
     */
    timestamp: bigint;
    /**
     * @generated from protobuf field: uint64 block_number = 4
     */
    block_number: bigint;
    /**
     * @generated from protobuf field: string event_key = 5
     */
    event_key: string;
}
/**
 * Request to query logs
 *
 * @generated from protobuf message torii.sinks.log.QueryLogsRequest
 */
export interface QueryLogsRequest {
    /**
     * @generated from protobuf field: uint32 limit = 1
     */
    limit: number; // Max number of logs to return (default: 5)
}
/**
 * Response with list of logs
 *
 * @generated from protobuf message torii.sinks.log.QueryLogsResponse
 */
export interface QueryLogsResponse {
    /**
     * @generated from protobuf field: repeated torii.sinks.log.LogEntry logs = 1
     */
    logs: LogEntry[];
}
/**
 * Log update for streaming
 *
 * @generated from protobuf message torii.sinks.log.LogUpdate
 */
export interface LogUpdate {
    /**
     * @generated from protobuf field: torii.sinks.log.LogEntry log = 1
     */
    log?: LogEntry;
    /**
     * @generated from protobuf field: int64 timestamp = 2
     */
    timestamp: bigint;
}
// @generated message type with reflection information, may provide speed optimized methods
class LogEntry$Type extends MessageType<LogEntry> {
    constructor() {
        super("torii.sinks.log.LogEntry", [
            { no: 1, name: "id", kind: "scalar", T: 4 /*ScalarType.UINT64*/, L: 0 /*LongType.BIGINT*/ },
            { no: 2, name: "message", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "timestamp", kind: "scalar", T: 3 /*ScalarType.INT64*/, L: 0 /*LongType.BIGINT*/ },
            { no: 4, name: "block_number", kind: "scalar", localName: "block_number", T: 4 /*ScalarType.UINT64*/, L: 0 /*LongType.BIGINT*/ },
            { no: 5, name: "event_key", kind: "scalar", localName: "event_key", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<LogEntry>): LogEntry {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.id = 0n;
        message.message = "";
        message.timestamp = 0n;
        message.block_number = 0n;
        message.event_key = "";
        if (value !== undefined)
            reflectionMergePartial<LogEntry>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: LogEntry): LogEntry {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* uint64 id */ 1:
                    message.id = reader.uint64().toBigInt();
                    break;
                case /* string message */ 2:
                    message.message = reader.string();
                    break;
                case /* int64 timestamp */ 3:
                    message.timestamp = reader.int64().toBigInt();
                    break;
                case /* uint64 block_number */ 4:
                    message.block_number = reader.uint64().toBigInt();
                    break;
                case /* string event_key */ 5:
                    message.event_key = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: LogEntry, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* uint64 id = 1; */
        if (message.id !== 0n)
            writer.tag(1, WireType.Varint).uint64(message.id);
        /* string message = 2; */
        if (message.message !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.message);
        /* int64 timestamp = 3; */
        if (message.timestamp !== 0n)
            writer.tag(3, WireType.Varint).int64(message.timestamp);
        /* uint64 block_number = 4; */
        if (message.block_number !== 0n)
            writer.tag(4, WireType.Varint).uint64(message.block_number);
        /* string event_key = 5; */
        if (message.event_key !== "")
            writer.tag(5, WireType.LengthDelimited).string(message.event_key);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message torii.sinks.log.LogEntry
 */
export const LogEntry = new LogEntry$Type();
// @generated message type with reflection information, may provide speed optimized methods
class QueryLogsRequest$Type extends MessageType<QueryLogsRequest> {
    constructor() {
        super("torii.sinks.log.QueryLogsRequest", [
            { no: 1, name: "limit", kind: "scalar", T: 13 /*ScalarType.UINT32*/ }
        ]);
    }
    create(value?: PartialMessage<QueryLogsRequest>): QueryLogsRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.limit = 0;
        if (value !== undefined)
            reflectionMergePartial<QueryLogsRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: QueryLogsRequest): QueryLogsRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* uint32 limit */ 1:
                    message.limit = reader.uint32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: QueryLogsRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* uint32 limit = 1; */
        if (message.limit !== 0)
            writer.tag(1, WireType.Varint).uint32(message.limit);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message torii.sinks.log.QueryLogsRequest
 */
export const QueryLogsRequest = new QueryLogsRequest$Type();
// ...
class LogUpdate$Type extends MessageType<LogUpdate> {
    constructor() {
        super("torii.sinks.log.LogUpdate", [
            { no: 1, name: "log", kind: "message", T: () => LogEntry },
            { no: 2, name: "timestamp", kind: "scalar", T: 3 /*ScalarType.INT64*/, L: 0 /*LongType.BIGINT*/ }
        ]);
    }
    create(value?: PartialMessage<LogUpdate>): LogUpdate {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.timestamp = 0n;
        if (value !== undefined)
            reflectionMergePartial<LogUpdate>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: LogUpdate): LogUpdate {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* torii.sinks.log.LogEntry log */ 1:
                    message.log = LogEntry.internalBinaryRead(reader, reader.uint32(), options, message.log);
                    break;
                case /* int64 timestamp */ 2:
                    message.timestamp = reader.int64().toBigInt();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: LogUpdate, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* torii.sinks.log.LogEntry log = 1; */
        if (message.log)
            LogEntry.internalBinaryWrite(message.log, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* int64 timestamp = 2; */
        if (message.timestamp !== 0n)
            writer.tag(2, WireType.Varint).int64(message.timestamp);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message torii.sinks.log.LogUpdate
 */
export const LogUpdate = new LogUpdate$Type();